Fix build on osx.
Some legitimate uninitialized variables caught by clang.

The molfile initialization was depending upon file-scope initializers
happening in a certain order; this is now fixed.

Use -Wno-error on darwin instead of -Wno-cpp.

Use iostream instead of printf in a few places where the compiler is
picky about the format string in a platform-dependent way.