dms export now avoids exceptions in destructors.
Previously, if a dms file were written to a bad path, an exception would
be thrown within the destructor handler passed to boost::shared_ptr.
This violates the policy  that the destructor must not throw.  Somehow
we've gotten away with this, but when I tried replacing boost:shared_ptr
with std::shared_ptr, the program just aborted instead of propagating
the exception.