Make dtr-validate failure cases be superset of d_validate
Committed to head some changes Justin added in version 113
that somehow are in the garden but not in head.

Added total_bytes method to FrameSetReader which is called
from python code to get the total size in bytes of frames
as reported by timekeys.

dtr-validate checks total_bytes() and compares against the
reported size of all frameXXXXXXXXX files to make sure that
the two are in agreement.

Compute delta_t in dtr-validate like d_validate used to
and assert an error is the change in delta_t between
frames is more than 0.02 (same magic value as d_validate)

Turned some previous molfile warnings into errors i.e.
a 0 byte frame will now be fatal.

Bump version to 114