Quickly return from TermTable::update_index if no work to be done
DESRESCode#2423

The expensive operation this commit is trying to avoid is the
initialization of a std::vector<bool> of size maxAtomId(). If i == n,
then the for-loop right below the initialization is never entered; and
the final line of the function, "_maxIndexId = n;" has no effect,
because i==n implies that _maxIndexId is already == n.