msys-core.
libmsys.so is now a symlink to a versioned shared library with a version
string derived from the git hash.

libmsys.so links to libmsys-core.so, which defines a function called
called msys::abi_version().  That version should remain stable for some
time, but we check its value against the value against which we were
compiled in the python extension.

Renamed desres-install-helper to garden-install-helper, since apparently
in CentOS7 the new garden tools require it.

Get the version from the install helper rather than version.py.

    Change-Id: I6bb2e95643314bb16f16bd1b7f4d9744e864f382

Change-Id: If085e426b8fe17233d5a262fb917b08cfff125e4