Make old, non-monotonic stks only generate warning.
Around two or more years ago we started throwing failure when a dtr in
an stk started at a time prior to any time in the preceding dtr.  But
some old anton1 trajectories still have this issue.  When this condition
is detected in dtrs that live in legacy storage (lockers), print a
warning instead of throwing failure.