msys.ConvertToRdkit fixed for RDKit 2017.09
Chem.DetectBondStereochemisty had a new precondition added to make
sure the conformer passed to it belonged to the molecule. The RDKit
Python API makes a copy of the Conformer object passed to
ROMol.AddConformer, so we need to grab the conformer again using
ROMol.GetConformer with the conformer index passed back to us.