Make unbuffered dms around 10x faster over NFS
Use pragma locking_mode=EXCLUSIVE to avoid taking file locks on each
disk access; instead hold the lock for the duration of the read.

Also use journal_mode=MEMORY.  This makes file corruption possible if
the process aborts midway, but we can patch this up with an atomic
rename.  It provides a significant speedup on top of the locking_mode
optimization.