fix for DESRESCode#4348
added the following routine:
SystemPtr CanonicalizeMoleculeByTopids(SystemPtr mol,
                                       IdList const& atoms,
                                       std::map<Id, Id> &aid_to_canId,
                                       std::map<Id, Id> &bid_to_canId);

it takes a molecule, and reorders the atoms so that all topologically
identical atoms are grouped together, and topological groups are sorted
by ascending id. The bonds are also sorted by the topological ids of their
constituent atoms.

updated the bondorderassigner to canonicalize the molecule before
calculating the formal charges and bond orders