Better handling of ct data with mixed types.
Previously, for a system with multiple cts, where the same key appears in
multiple cts but with different types, writing to a dms file would fail.
We now support writing such systems.

However, in order to read these files and maintain backward compatibility,
we need to make a small compromise.  Existing versions of msys expect
a declared type for every column in a table.  However, when there is
a declared type, sqlite will automatically convert your values to that
type, and not tell you what the original type was.  We therefore try to
convert the value to a numeric type using sqlite's own functions; this
is close to what msys does anyway with data in sdf files.  Thus, if you
manually add a ct property whose declared type is a string but whose value
"looks" numeric, you'll get a numeric type back from the dms file.

DESRESCode#4414.

Change-Id: I73118c2185c66a427b419b39f08e69f0ad70abe7