Bring back the python2 build; for now.
VMD is still on python2, but we want to upgrade the zendo version used
by VMD in order to use the filtered trajectory API.  But zendo uses a
newer version of msys in order to support compressed trajectories, and
that new version dropped python2 support.  For some reason, mixing the
python versions in the build is causing problems for zendo.  Porting VMD
to python3 may take a while, so keep python2 around in msys until that
can be done.

This reverts commit 3bdacc6941b8285b5afe2c9b4d642d48542dc8e5.

Change-Id: Iee921dc2cb4bc337129681c96ffcfef1cdd0741a