Replace mae string storage with a simple registry.
The old code was trying be space efficient by using a single contiguous
storage block for strings, and by representing those strings with an
custom encoding scheme.  Unfortunately it also had O(n^2) insertion
time for new values, which was somehow tolerable when positions in mae
files were truncated to some small number of decimals but became a big
performance problem when we stopped truncating in version 1.7.313.

DESRESCode#4777