More work on wheel compatibility.
At this point I _think_ the only blocker is the fact that our
boost_iostreams links liblzma, which isn't in the whitelist for any of
the 2_28 or 2_31 wheel versions and which might have GPL code in it so
we probably couldn't distribute it anyway.  We don't actually use
liblzma in boost so at some point it might be nice to try building a new
boost without that dependency and see if we can get further.